server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "The backend service is a Django REST framework to communicate with Azure OpenAI GPT-4o model."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.4.0",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "loguru>=0.7.2",
    "openai>=1.52.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "add-trailing-comma>=4.0.0",
    "coverage>=7.6.7",
    "crosszip>=1.0.0",
    "django-stubs>=5.1.0",
    "djangorestframework-stubs>=3.15.1",
    "locust>=2.32.0",
    "prek>=0.3.10",
    "pytest-django>=4.9.0",
    "pytest-random-order>=1.1.1",
    "pytest>=8.3.3",
    "ruff>=0.15.11",
    "ty>=0.0.32",
]

[tool.uv]
required-version = ">=0.11.2"

[tool.ruff]
fix = true
preview = true
unsafe-fixes = true
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "CPY", "D100", "D203", "D212", "DOC502"]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = [
    # docstring requirements
    "D",
    "DOC",
    # tests can have asserts
    "S101",
    # magic values in assertions are fine
    "PLR2004",
    # parametrized tests naturally have many arguments
    "PLR0913",
    "PLR0917",
    # fixtures used for side effects don't need to be referenced
    "ARG001",
    # `== ""` assertions are intentionally explicit
    "PLC1901",
]

[tool.ruff.format]
preview = true
docstring-code-format = true

[tool.ty.src]
exclude = [".venv", "build", "dist"]

[tool.ty.environment]
python-version = "3.13"

[tool.coverage.run]
omit = [
    # boilerplate generated by Django
    "chatgptserver/chatgptserver/*",
    "chatgptserver/manage.py",
    # no need to check openai implementation
    "chatgptserver/api/utils.py",
]

[tool.pytest.ini_options]
addopts = [
    # error on using unregistered marker
    "--strict-markers",
    # show extra test summary info for everything
    "-ra",
]
xfail_strict = true
python_files = ["test_*.py", "test-*.py", "tests.py", "test.py"]
